requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
fredapi==0.5.2
requests==2.32.3
scikit-learn==1.5.1
pyportfolioopt==1.5.6
matplotlib==3.8.4
plotly==5.24.1
duckdb==1.1.1